// src/app/models/pages.ts
export interface PageResponse {
  kind: string
  items: Page[]
  etag: string
}

export interface Page {
  kind: string;
  id: string;
  blog: {
    id: string;
  };
  published: string;
  updated: string;
  url: string;
  selfLink: string;
  title: string;
  content: string;
  author: {
    id: string;
    displayName: string;
    url: string;
    image: {
      url: string;
    };
  };
  etag: string;
}